+++ src/pybricks/__init__.py
@@ -1,0 +1,8 @@
+from typing import Tuple
+
+
+version: Tuple[str, str, str] = (
+    "hub",
+    "3.2.0b1",
+    "v3.2.0b1-GIT_HASH on DATE",
+)